Fix randmFillSync
